Many changes ... again.
I must learn to use small commits.
This commit will not work. Changes:

- Bone deformation in shader use a 4x4 affine transformation matrix
- Correct physics environment deinitialization and tweaked gravity
- `XMTRANSFORM` does not use a `XMMATRIX` for its storage anymore (it
does not store the rotation origin)
- Converted `PMX::Bone` into an interface and created two classes:
`detail::RootBone` and `detail::BoneImpl`
- Debug rendering for joints
- Fixed the wierd rendering error by disabling the OM stage for the
geometry shader

+++ XBeat/Renderer/Texture.cpp
@@ -68,7 +68,7 @@ bool Texture::Initialize(ID3D11Device *device, const std::wstring &file)
 	char s[1024];
 	size_t len;
 	wcstombs_s<1024>(&len, s, file.c_str(), 1024);
-	texture->SetPrivateData(WKPDID_D3DDebugObjectName, len, s);
+	texture->SetPrivateData(WKPDID_D3DDebugObjectName, (UINT)len, s);
 #endif
 
 	return true;
